メソッド抽出。#12
